refactor: Extract fn from client/mobile to helpers
